modified:   Talleres/Taller 5/PVBI_Taller5_1_932.c 	new file:   Talleres/Taller 5/a.exe 	modified:   Talleres/Taller 5/anotherfile.c 	new file:   Talleres/Taller 5/output/anotherfile.exe

+++ Talleres/Taller 5/anotherfile.c
@@ -1,1 +1,2 @@
-double saldo = 100.0;
+
+int saldo = 100.0;
